Moved the RuptureGetter inside getters [skip hazardib]
