+++ openquake/calculators/event_based.py
@@ -35,9 +35,9 @@ from openquake.hazardlib.probability_map import ProbabilityMap
 from openquake.hazardlib.stats import compute_pmap_stats
 from openquake.risklib.riskinput import str2rsi, rsi2str, indices_dt
 from openquake.baselib import parallel
-from openquake.commonlib import calc, util, readinput, source
+from openquake.commonlib import calc, util, readinput
 from openquake.calculators import base
-from openquake.calculators.getters import GmfGetter
+from openquake.calculators.getters import GmfGetter, EBRupture
 from openquake.calculators.classical import (
     ClassicalCalculator, saving_sources_by_task)
 
@@ -173,8 +173,8 @@ def _build_eb_ruptures(
                 # set a bit later, in set_eids
                 events.append((0, src.src_group_id, ses_idx, sampleid))
         if events:
-            yield calc.EBRupture(
-                rup, indices, numpy.array(events, calc.event_dt), serial)
+            yield EBRupture(rup, indices, numpy.array(events, calc.event_dt),
+                            serial)
 
 
 def get_events(ebruptures):
